KONTRON: gpio: pca953x: Disable wake-up support
This is a workaround to mitigate crash on resume.

On suspend, non-wakeup IRQ lines are masked:

  suspend_enter() -> dpm_suspend_noirq() -> ... -> suspend_device_irq()

Wake-up IRQ is marked as such by irq_chip->irq_set_wake(irq_data, 1).
Devices that supports wake-up signal it to IRQ subsystem by
enable_irq_wake(irq) in their ->suspend() PM callback. This in turns
calls irq_chip->irq_set_wake(irq_data, 1) on all parent irq_chips.

gpio-mxc's irq_chip supports irq_set_wake. It also supports additional
non-standard wake-up mechanism where wake-up pins are configured by
pad-wakeup device tree property, and this configuration is communicated
to SCU via RPC in ->suspend() PM callback. In this latter case
gpio-mxc's IRQ lines are not marked as wake-up IRQs and Linux masks them
during suspend.

In our case (sAMX8X), gpio-mxc's IRQ line is marked as wake-up IRQ
because gpio-charger marks itself as wake-up source by calling
enable_irq_wake(). This was not the case in 4.14.98 because pca953x
didn't support power management and so set_irq_wake_real() stopped on
it. However, in 5.4.3 it gained that feature.

When the LID triggers resume, the IRQ handler of gpio-mxc,
mx3_gpio_irq_handler(), is executed right after

  suspend_enter() -> arch_suspend_enable_irqs()

The handler then reads something from GPIO register map but this is
virtual memory which at this point is not yet mapped, and the access
results in synchronous external abort, a carsh. See log at the bottom.

It is not yet clear to me whether wake-up IRQ handlers are supposed to
be run this soon in resume process. So as a workaround this commit makes
pca953x wake-up non-capable just as it was in 4.14.98.

  PM: Timekeeping suspended for 18.901 seconds
  Internal error: synchronous external abort: 96000210 [#1] PREEMPT SMP
  Modules linked in:
  CPU: 0 PID: 560 Comm: sleep Not tainted 5.4.3-00282-g3e018c99f6da #284
  Hardware name: Kontron SMARC-sAMX8X QuadXPlus (DT)
  pstate: a0000085 (NzCv daIf -PAN -UAO)
  pc : mx3_gpio_irq_handler+0x70/0xe0
  lr : generic_handle_irq+0x34/0x50
  sp : ffff800010003f40
  x29: ffff800010003f40 x28: ffff000835f28e80
  x27: ffff8000120b8000 x26: ffff800010004000
  x25: ffff800010000000 x24: ffff000838016200
  x23: 0000000000000000 x22: ffff800012ffb900
  x21: ffff0008399d1880 x20: ffff000838650658
  x19: ffff8000121406d8 x18: 0000000000000010
  x17: 0000000000000000 x16: 0000000000000000
  x15: ffffffffffffffff x14: 0000000000000000
  x13: 003d090000000000 x12: 00003d0900000000
  x11: 0000000000000000 x10: 0000000000000040
  x9 : ffff8000120d8300 x8 : ffff8000120d82f8
  x7 : ffff0008360b3c10 x6 : 0000000000000000
  x5 : ffff0008360b3b50 x4 : 0000000000000000
  x3 : 0000000000000000 x2 : ffff80001243d020
  x1 : 0000000000000018 x0 : ffff8000251b0018
  Call trace:
   mx3_gpio_irq_handler+0x70/0xe0
   generic_handle_irq+0x34/0x50
   __handle_domain_irq+0x94/0xc0
   gic_handle_irq+0xc4/0x158
   el1_irq+0xbc/0x180
   preempt_schedule_irq+0xa0/0x130
   arm64_preempt_schedule_irq+0x18/0x20
   el1_irq+0xd4/0x180
   arch_suspend_enable_irqs+0x1c/0x28
   suspend_devices_and_enter+0x4f0/0x8c8
   pm_suspend+0x1e4/0x418
   state_store+0xd0/0x108
   kobj_attr_store+0x14/0x28
   sysfs_kf_write+0x68/0x78
   kernfs_fop_write+0x160/0x1f8
   __vfs_write+0x48/0x80
   vfs_write+0xd8/0x1b0
   ksys_write+0x74/0xf0
   __arm64_sys_write+0x24/0x30
   el0_svc_common.constprop.2+0xf4/0x170
   el0_svc_handler+0x74/0x90
   el0_svc+0x8/0xc
  Code: f9400aa0 f9400841 b9401821 8b010000 (b9400000)
  ---[ end trace e85573b93fe9ebda ]---
  Kernel panic - not syncing: Fatal exception in interrupt
  Kernel Offset: disabled
  CPU features: 0x0002,20002008
  Memory Limit: none
  ---[ end Kernel panic - not syncing: Fatal exception in interrupt ]---