Add mpcflex-p1 DTS + modify kontron-samx8x.dtsi
A new separate DTS file has been created named mpcflex-p1.dts

It replaces the main DTS that is used normally in the Kontron BSP
(kontron-samx8x-qxp.dts).

Kontron-samx8x-qxp.dts includes (builds on top of) kontron-samx8x.dtsi
as is customary.

It ought to be possible to create a new hardware description by only
building on top of the .dtsi definition, just like kontron-samx8x-qxp
does.

However, some of the kontron-samx8x.dtsi refers also to things
defined in the EVAL board file: smarc-evaluation-carrier-v2.dtsi
and kontron-samx8x-qxp.dts also #includes the smarc-evaluation DTS file.

The purpose of this rework is ultimately to remove anything that is
unique to the evaluation board, since it is not relevant for our custom
hardware and to define or redefine as required by our hardware.

Certain definitions similar to smarc-evaluation-carrier-v2 have been
copied into mpcflex-p1.dts, because there are references in the original
kontron-samx8x.dtsi that need to be fulfilled and that were previously
fulfilled only in the smarc evaluation tree.  (After including a few
things directly in mpcflex-p1.dts, there is no need to include
smarc evaluation DTS anymore.) Some of those definitions may be later
redefined or tweaked, if further differences are identified between the
EVAL board and our hardware.

Several comments were added to kontron-samx8x.dtsi to explain each
node. This is only for our own future development convenience.

The LVDS display panel that is described in kontron-samx8x.dtsi seems
out of place.  As a starting point it is being modified here, but
it seems more that it ought to be in the board-specific file rather than
the more generic one, so it might move later on.

The sound chip, wm8904, of the eval board has been removed.  Later we
will add our own sound chip.

Not complete, but this starts the process of simplifying down the device
tree to match what is _actually_ in the custom hardware, as opposed to
the eval board.