flexcan: TRY disable xceiver regulator test
There is some error in the voltage regulator definition and check for
the CAN tranceiver.

We know that one difference is that there is no enable-signal for our
external CAN tranceiver (it is constantly on).  This enable signal was
however used in the eval board's CAN setup.  That might affect the check
for the voltage regulator?

A first resolution that works is to not check the regulator, so that we
can get the interfaces up and start testing them.