Redefine kontron-samx8x-qxp.dts (development hack)
This commit fakes the content of kontron-samx8x-qxp and inserts our own
hardware definition instead (equivalent to mpcflex-p1.dts).

Explanation:

At this point we have our own hardware definition (Device Tree) in
mpcflex-p1.dts.  The bootloader has been programmed to load
kontron-samx8x-qxp.dts by default.

It *is* possible to stop the bootloader and specify it to load our DT
instead before booting the kernel, but we don't have serial console on
all hardwares so it is inconvenient for testing on those hardwares.

Long term this would be confusing but until it is time to do the proper
thing and reprogram the bootloader to load the DTB with the correct file
name -- for now kontron-samx8x-qxp will contain our custom hardware
description instead.