kontron-samx8x.dtsi: Restore old LVDS settings + add comments 2
This restores settings to something that works.  It is basically
reusing an AUO panel setting, even if it is not an AUO panel,
using default settings for many parameters and and then adding our own
exact resolution + timing.  This is a working configuration but
setting a few more parameters later might remove some warning printouts
during boot.