Device tree parameters for 1024x768 LVDS panel - final(?) adjustment
This is a combination of several attempts to get the settings right.
In the end, the panel timing was set according to data sheet
information for the chosen display, but some other settings (for the
LVDS bridge chip) were reverted back to being unset, or set to what
the driver had as default because that works correctly.

Apart from panel timing, noteworthy explicit settings are:
vesa-24 coding, and 8 bit color depth per channel (=24 bit)

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>