Device tree parameters for 1024x768 LVDS panel - take 2
This is a combination of several attempts to get the settings right.
In the end, the panel timing was set according to data sheet
information for the chosen display, but many other settings (for the
LVDS bridge chip) were reverted back to being unset - i.e. many of the
defaults work correctly.

In particular this means there is a dual-channel LVDS setting for the
bridge, but the system only uses one channel.  We might look into that
later on.

Apart from panel timing, noteworthy explicit settings that remain are:
vesa-24 coding, and 8 bit color depth per channel (=24 bit)

Signed-off-by: Gunnar Andersson <gunnar_dev@nospam@novaspring.eu>